# internal/database/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.29.0

package database

import (
	"database/sql"
	"time"

	"github.com/google/uuid"
)

type Email struct {
	TransactionUuid    uuid.UUID
	Consumer           string
	UserName           string
	EmailSubject       sql.NullString
	EmailText          sql.NullString
	Html               sql.NullString
	SenderName         string
	SenderEmail        string
	RecipientsName     string
	RecipientsEmail    string
	Status             string
	CreatedAt          time.Time
	EnableOpenTracking bool
	Opened             bool
}

type TextMessage struct {
	MessageID uuid.UUID
	Consumer  string
	UserName  string
	Sender    string
	Recipient string
	Status    string
	ApiKey    string
	ErrCode   sql.NullString
	Price     string
	CreatedAt time.Time
}